Fixed compiler warning.
git-svn-id: svn+ssh://svn.code.sf.net/p/supertuxkart/code/main/trunk@8098 178a84e3-b1eb-0310-8ba1-8eac791a3b58

+++ src/guiengine/widgets/label_widget.cpp
@@ -143,7 +143,7 @@ void LabelWidget::setText(const wchar_t *text, bool expandIfNeeded)
     if (m_scroll_speed > 0)
     {
         //m_scroll_offset = (float)m_element->getAbsolutePosition().getWidth();
-        m_scroll_offset = m_w;
+        m_scroll_offset = (float)m_w;
     }
     
     Widget::setText(text);
